+++ src/main/java/userinputhistory.txt
@@ -1,4 +1,3 @@
-[T][ ] newTask
-[E][ ] e1 (at: 2022-04-05)
-[E][ ] 32 (at: 2021-04-02)
-[D][✓]  e1 
+[T][ ] home
+[T][ ] buy milk
+[E][✓]  32
